+++ lockFileLint.js
@@ -12,7 +12,6 @@
  * under the License.
  */
 
-
 const { spawnSync } = require('child_process');
 const path = require('path');
 const fs = require('fs');