+++ packages/generator-one-app-module/generators/app/templates/base-module/src/appConfig.js
@@ -1,7 +1,0 @@
-import csp from './csp';
-
-// Read about appConfig:
-// https://github.com/americanexpress/one-app/blob/master/docs/api/modules/App-Configuration.md
-export default {
-  csp,
-};