+++ packages/one-app-bundler/webpack/module/webpack.client.js
@@ -96,5 +96,5 @@ module.exports = (babelEnv) => {
         }),
       ],
     }
-  ));
+  ), 'client');
 };